Moving result from function based view to class based view.
Adding new mix in for authorisation checks.  This is to allow a more flexible approach to authorisation in the future.

Also update template to use context object for data.  UX improvement to warn user if no result have yet been recorded against a temp.

Adding coveralls for a published view of code coverage pushed out form build.

Adding coverall coverage badge for quick coverage and link to metrics on coverall.